commited

+++ database/migrations/05_projects_outcomes.ts
@@ -1,25 +1,24 @@
-import { BaseSchema } from '@adonisjs/lucid/schema'
+import { BaseSchema } from '@adonisjs/lucid/schema';
 
-export default class ProjectOutcomes extends BaseSchema {
-  protected tableName = 'project_outcomes';
+export default class ProjectsOutcomes extends BaseSchema {
+  protected tableName = 'projects_outcomes';
 
   public async up() {
-    await this.schema.createTable(this.tableName, (table) => {
-      table.increments('id').primary();
-      table.string('verb').notNullable();
-      table.string('subject').notNullable();
-      table.string('from_to').notNullable(); // Corrected column name
-      table.string('units').notNullable();
-      table.string('units_of_measure').notNullable();
-
-      // Corrected syntax for defining timestamps with timezone
-      table.timestamp('created_at', { useTz: true });
-      table.timestamp('updated_at', { useTz: true });
-    });
+    if (!this.schema.hasTable(this.tableName)) {
+      await this.schema.createTable(this.tableName, (table) => {
+        table.increments('id').primary();
+        table.string('verb').notNullable();
+        table.string('subject').notNullable();
+        table.string('from_to').notNullable(); // Corrected column name
+        table.string('units').notNullable();
+        table.string('units_of_measure').notNullable();
+        table.timestamps(); // Define timestamps without timezone
+      });
+    }
   }
-
   public async down() {
-    await this.schema.dropTable(this.tableName);
+    if (!this.schema.hasTable(this.tableName)) {
+      await this.schema.dropTable(this.tableName);
+    }
   }
 }
-
